+++ src/cli/commands/self_install.rs
@@ -1,6 +1,8 @@
 use crate::cli::{version::update_bin_exe, HOME_DIR};
+use anyhow::Context;
 use clap::Args;
 use colored::Colorize;
+use std::env::current_exe;
 #[derive(Debug, Args)]
 pub struct SelfInstallCommand {
     /// Skip adding the bin directory to the PATH
@@ -68,7 +70,7 @@ and then restart your shell.
             );
         }
 
-        update_bin_exe().await?;
+        update_bin_exe(&current_exe().context("failed to get current exe path")?).await?;
 
         Ok(())
     }